[update] : scrumboard supabase type 제외 변경
