+++ nbc-nep-front/src/components/scrumboard/ScrumBoardHeader.tsx
@@ -1,19 +1,19 @@
 import { StLine } from "@/components/spaces/SpaceListHeader";
-import styled from "styled-components";
 import SpaceSearchForm from "@/components/spaces/SpaceSearchForm";
+import {
+  StHeaderWrapper,
+  StSortTitleContainer,
+} from "./styles/scrumBoardHeader.styles";
 
 export default function ScrumBoardHeader() {
   return (
     <StHeaderWrapper>
       <StSortTitleContainer>
         <span>최근 순</span>
-        <StLine></StLine>
+        <StLine />
         <span>이름 순</span>
       </StSortTitleContainer>
       <SpaceSearchForm />
     </StHeaderWrapper>
   );
 }
-
-const StHeaderWrapper = styled.div``;
-const StSortTitleContainer = styled.div``;
