+++ nbc-nep-front/src/components/scrumboard/ScrumBoardList.tsx
@@ -1,9 +1,9 @@
+import ScrumBoardHeader from "@/components/scrumboard/ScrumBoardHeader";
 import { useGetUserSpaces } from "@/hooks/query/useSupabase";
 import useAuthStore from "@/zustand/authStore";
 import { StCardListWrapper } from "../spaces/SpaceList";
 import ScrumBoardCard from "./ScrumBoardCard";
-import styled from "styled-components";
-import ScrumBoardHeader from "@/components/scrumboard/ScrumBoardHeader";
+import { StScrumBoardListWrapper } from "./styles/scrumBoardList.styles";
 
 export default function ScrumBoardList() {
   const user = useAuthStore.use.user();
@@ -19,9 +19,3 @@ export default function ScrumBoardList() {
     </StCardListWrapper>
   );
 }
-
-const StScrumBoardListWrapper = styled.ul`
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  gap: ${(props) => props.theme.spacing[24]};
-`;
