+++ nbc-nep-front/src/components/scrumboard/constants.ts
