+++ nbc-nep-front/src/components/scrumboard/detail/ScrumBoard.tsx
@@ -1,4 +1,3 @@
-import { StCTAButton } from "@/components/common/button/button.styles";
 import ModalPortal from "@/components/modal/ModalPortal";
 import CreateCategoryModal from "@/components/modal/scrumboardModal/CreateCategoryModal";
 import CreateBackDrop from "@/components/scrumboard/detail/CreateBackDrop";
@@ -13,22 +12,28 @@ import useScrumBoardItemBackDropStore from "@/zustand/createScrumBoardItemStore"
 import { AnimatePresence } from "framer-motion";
 import { useParams } from "next/navigation";
 import { WheelEvent, useEffect } from "react";
-import styled from "styled-components";
+import {
+  StAddCategoryBtn,
+  StScrumBoardContainer,
+  StScrumBoardWrapper,
+} from "../styles/scrumBoard.styles";
 import ScrumBoardCategory from "./ScrumBoardCategory";
 import ScrumBoardHeader from "./ScrumBoardHeader";
 
 export default function ScrumBoard() {
-  const { space_id } = useParams();
-  const spaceId = space_id as string;
+  const { space_id: spaceId } = useParams();
   const { openCreateCategoryModal, isCreateCategoryModalOpen } = useModal();
   const { setCategories } = useScrumBoard();
-  const categories = useGetCategories(spaceId);
+  const categories = useGetCategories(spaceId as string);
   const isCreateBackDropOpen = useScrumBoardItemBackDropStore.use.isOpen();
-  const spaceData = useGetSpaceQuery(spaceId);
+  const spaceData = useGetSpaceQuery(spaceId as string);
 
-  useCategorySubscribe(spaceId);
+  useCategorySubscribe(spaceId as string);
   // items에 대한 구독 커스텀훅
-  useScrumBardItemsSubscribe(spaceId, categories as Kanban_categories[]);
+  useScrumBardItemsSubscribe(
+    spaceId as string,
+    categories as Kanban_categories[]
+  );
 
   useEffect(() => {
     setCategories(categories!);
@@ -41,14 +46,14 @@ export default function ScrumBoard() {
   const [handleFocus, handleBlur] = useFocusInput();
 
   const handleWheel = (e: WheelEvent<HTMLDivElement>) => {
-    if (e.deltaY != 0) {
+    if (e.deltaY !== 0) {
       e.currentTarget.scrollLeft += e.deltaY;
     }
   };
 
   return (
     <StScrumBoardWrapper>
-      <ScrumBoardHeader title={spaceData?.title!} />
+      <ScrumBoardHeader title={spaceData ? spaceData.title : ""} />
       <AnimatePresence>
         <StScrumBoardContainer onWheel={handleWheel}>
           <div onFocus={handleFocus} onBlur={handleBlur}>
@@ -74,29 +79,3 @@ export default function ScrumBoard() {
     </StScrumBoardWrapper>
   );
 }
-
-const StScrumBoardWrapper = styled.div`
-  position: relative;
-`;
-
-const StScrumBoardContainer = styled.div`
-  max-width: 1200px;
-  width: 100%;
-  overflow: auto;
-  margin: 0 auto;
-  position: relative;
-  padding: 0 ${(props) => props.theme.spacing[24]};
-  > div {
-    display: flex;
-    justify-content: flex-start;
-    align-items: flex-start;
-    gap: ${(props) => props.theme.spacing[12]};
-    position: relative;
-  }
-`;
-
-const StAddCategoryBtn = styled(StCTAButton)`
-  display: block;
-  width: 320px;
-  height: ${(props) => props.theme.unit[80]};
-`;
