+++ nbc-nep-front/src/components/scrumboard/detail/ScrumBoardCategory.tsx
@@ -1,6 +1,5 @@
 import NoContents from "@/components/common/NoContents";
-import { StCTAButton } from "@/components/common/button/button.styles";
-import { BACK_DROP_TYPE_CREATE } from "@/components/scrumboard/constants/constants";
+import { BACK_DROP_TYPE_CREATE } from "@/components/scrumboard/constants";
 import { useGetCategoryItems } from "@/hooks/query/useSupabase";
 import useDropItem from "@/hooks/scrumBoard/useDropItem";
 import { fadeInOut } from "@/styles/animations";
@@ -8,7 +7,11 @@ import { Kanban_categories } from "@/types/supabase.tables.types";
 import useScrumBoardItemBackDropStore from "@/zustand/createScrumBoardItemStore";
 import { AnimatePresence, motion } from "framer-motion";
 import { WheelEvent } from "react";
-import styled from "styled-components";
+import {
+  StAddItemBtn,
+  StCategoryWrapper,
+  StItemsContainer,
+} from "../styles/scrumBoardCategory.styles";
 import CategoryHeader from "./CategoryHeader";
 import ScrumBoardItem from "./ScrumBoardItem";
 
@@ -45,10 +48,14 @@ export default function ScrumBoardCategory({ category }: Props) {
       {items?.length ? (
         <StItemsContainer ref={drop} onWheel={handleWheel}>
           <AnimatePresence>
-            {items?.map((item, index) => {
+            {items?.map((item) => {
               return (
-                <motion.div key={index} {...fadeInOut({ y: 5 })}>
-                  <ScrumBoardItem key={index} item={item} category={category} />
+                <motion.div key={item.id} {...fadeInOut({ y: 5 })}>
+                  <ScrumBoardItem
+                    key={`${item.id}scrumboardItem`}
+                    item={item}
+                    category={category}
+                  />
                 </motion.div>
               );
             })}
@@ -66,52 +73,3 @@ export default function ScrumBoardCategory({ category }: Props) {
     </StCategoryWrapper>
   );
 }
-
-const StCategoryWrapper = styled(motion.div)<{ $isOver: boolean }>`
-  // 임의로 설정한 너비
-  min-width: 384px;
-  height: 100%;
-  box-sizing: content-box;
-  background-color: ${(props) => props.theme.color.bg.secondary};
-  padding: ${(props) => props.theme.spacing[16]};
-  padding-top: 0;
-  ${(props) =>
-    props.$isOver
-      ? `border: 2px solid ${props.theme.color.border.interactive.primary}`
-      : `border : 2px solid ${props.theme.color.border.secondary}`};
-  transition: border 0.2s ease-in-out;
-  border-radius: ${(props) => props.theme.border.radius[12]};
-`;
-
-const StItemsContainer = styled.ul`
-  // 임의로 설정한 높이
-  height: calc(92vh - 320px);
-  display: flex;
-  flex-direction: column;
-  gap: ${(props) => props.theme.spacing[12]};
-  background-color: #f5f5f5;
-  border-radius: ${(props) => props.theme.border.radius[8]};
-  overflow: scroll;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-`;
-
-const StAddItemBtn = styled(StCTAButton)`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-top: ${(props) => props.theme.spacing[12]};
-  gap: ${(props) => props.theme.spacing[8]};
-  font-size: ${(props) => props.theme.heading.desktop.sm.fontSize};
-  & > span {
-    display: block;
-    width: ${(props) => props.theme.unit[14]};
-    height: ${(props) => props.theme.unit[14]};
-    margin-top: -${(props) => props.theme.unit[4]};
-    background: url("/assets/additem.svg") no-repeat center center;
-    background-size: contain;
-  }
-`;
