+++ nbc-nep-front/src/components/scrumboard/detail/ScrumBoardHeader.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import { StScrumBoardHeader } from "../styles/scrumBoardHeader.styles";
 
 interface Props {
   title: string;
@@ -10,26 +10,3 @@ export default function ScrumBoardHeader({ title }: Props) {
     </StScrumBoardHeader>
   );
 }
-
-const StScrumBoardHeader = styled.div`
-  max-width: 1200px;
-  width: 100%;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: start;
-  padding: ${(props) => props.theme.spacing[24]};
-  padding-top: ${(props) => props.theme.spacing[16]};
-  h1 {
-    font-family: var(--point-font);
-    font-weight: ${(props) => props.theme.heading.desktop.lg.fontWeight};
-    font-size: ${(props) => props.theme.heading.desktop["2xl"].fontSize};
-    color: ${(props) => props.theme.color.text.brand};
-  }
-  p {
-    font-family: var(--main-font);
-    font-weight: ${(props) => props.theme.body.lg.regular.fontWeight};
-    font-size: ${(props) => props.theme.body.lg.regular.fontSize};
-  }
-`;
