+++ nbc-nep-front/src/components/scrumboard/detail/ScrumBoardItem.tsx
@@ -1,5 +1,5 @@
 import MetaAvatar from "@/components/metaverse/avatar/MetaAvatar";
-import { BACK_DROP_TYPE_DETAIL } from "@/components/scrumboard/constants/constants";
+import { BACK_DROP_TYPE_DETAIL } from "@/components/scrumboard/constants";
 import useDragItem from "@/hooks/scrumBoard/useDragItem";
 import {
   GetKanbanItemsByAssignees,
@@ -7,7 +7,12 @@ import {
 } from "@/types/supabase.tables.types";
 import useScrumBoardItemBackDropStore from "@/zustand/createScrumBoardItemStore";
 import React from "react";
-import styled from "styled-components";
+import {
+  StAssigneesWrapper,
+  StListItem,
+  StMetaAvatarWrapper,
+  StUserInfoWrapper,
+} from "../styles/scrumBoardItem.styles";
 
 interface Props {
   item: GetKanbanItemsByAssignees;
@@ -16,10 +21,11 @@ interface Props {
 
 function ScrumBoardItem({ category, item }: Props) {
   const setIsOpen = useScrumBoardItemBackDropStore.use.setIsOpen();
-  const handleOpenItemDetail = (item: GetKanbanItemsByAssignees) => {
-    setIsOpen(category, item, BACK_DROP_TYPE_DETAIL);
+
+  const handleOpenItemDetail = (kanbanItem: GetKanbanItemsByAssignees) => {
+    setIsOpen(category, kanbanItem, BACK_DROP_TYPE_DETAIL);
   };
-  const { drag, didDrop, targetCategoryId, isDragging } = useDragItem(item);
+  const { drag, isDragging } = useDragItem(item);
 
   return (
     <StListItem
@@ -36,7 +42,10 @@ function ScrumBoardItem({ category, item }: Props) {
           <StAssigneesWrapper>
             {item.assignees.map((assignee, index) => {
               return (
-                <StMetaAvatarWrapper $index={index} key={index}>
+                <StMetaAvatarWrapper
+                  $index={index}
+                  key={assignee.userId + assignee.assigneesId}
+                >
                   <MetaAvatar
                     spaceAvatar={assignee.spaceAvatar}
                     width={24}
@@ -53,69 +62,5 @@ function ScrumBoardItem({ category, item }: Props) {
     </StListItem>
   );
 }
-
-const StListItem = styled.li<{ $isDragging: boolean }>`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: ${(props) => props.theme.spacing[4]};
-  align-self: stretch;
-  padding: ${(props) => props.theme.spacing[12]};
-  margin: 0;
-  width: 100%;
-  height: auto;
-  border-radius: ${(props) => props.theme.border.radius[12]};
-  ${(props) =>
-    props.$isDragging
-      ? `border: 2px solid ${props.theme.color.border.interactive.primary}`
-      : `border : 1px solid ${props.theme.color.border.secondary}`};
-  background-color: ${(props) => props.theme.color.bg.primary};
-  cursor: grab;
-
-  > p {
-    color: ${(props) => props.theme.color.text.secondary};
-    text-overflow: ellipsis;
-    font-size: ${(props) => props.theme.unit[14]};
-    font-style: normal;
-    font-weight: 400;
-    line-height: 150%; /* 19.5px */
-    letter-spacing: -0.26px;
-    font-family: var(--main-font);
-  }
-`;
-
-const StUserInfoWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  > div {
-    display: flex;
-    align-items: center;
-    height: 3rem;
-    > p {
-      flex: 1;
-    }
-  }
-`;
-
-const StAvatar = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-`;
-
-const StAssigneesWrapper = styled(StAvatar)`
-  position: relative;
-  height: 3rem;
-  flex: 1;
-`;
-
-const StMetaAvatarWrapper = styled.div<{ $index: number }>`
-  position: absolute;
-  right: ${(props) => props.$index * 15}px;
-  z-index: ${(props) => 1000 - props.$index};
-`;
 
 export default React.memo(ScrumBoardItem);
