+++ nbc-nep-front/src/components/scrumboard/detail/createBackDrop/CreateBackDropCTAButton.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import { StCreateBackDropCtaButton } from "../../styles/backdrop.styles";
 
 interface Props {
   buttonType: "button" | "submit" | "reset";
@@ -17,18 +17,3 @@ export default function CreateBackDropCtaButton({
     </StCreateBackDropCtaButton>
   );
 }
-
-const StCreateBackDropCtaButton = styled.button`
-  border: none;
-  font-size: ${(props) => props.theme.unit[14]};
-  font-family: var(--point-font);
-  font-weight: 300;
-  padding: ${(props) => props.theme.spacing[8]} 0;
-  color: ${(props) => props.theme.color.text.primary};
-  opacity: 0.3;
-  &:hover {
-    background-color: unset;
-    color: ${(props) => props.theme.color.text.primary};
-    opacity: 1;
-  }
-`;
