+++ nbc-nep-front/src/components/scrumboard/detail/createBackDrop/CreateBackDropTitle.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import { StCreateBackDropTitle } from "../../styles/backdrop.styles";
 
 interface Props {
   title: string;
@@ -7,12 +7,3 @@ interface Props {
 export default function CreateBackDropTitle({ title }: Props) {
   return <StCreateBackDropTitle>{title}</StCreateBackDropTitle>;
 }
-
-const StCreateBackDropTitle = styled.h2`
-  color: ${(props) => props.theme.color.text.primary};
-  font-family: var(--sub-font);
-  font-size: ${(props) => props.theme.unit[14]};
-  font-style: normal;
-  font-weight: 400;
-  line-height: 100%;
-`;
