+++ nbc-nep-front/src/components/scrumboard/detail/createBackDrop/CreateCategoryBackDrop.tsx
@@ -1,11 +1,14 @@
 import { updateCategoryItem } from "@/api/supabase/scrumBoard";
-import { BACK_DROP_TYPE_CREATE } from "@/components/scrumboard/constants/constants";
+import { BACK_DROP_TYPE_CREATE } from "@/components/scrumboard/constants";
 import useScrumBoard from "@/hooks/scrumBoard/useScrumBoard";
 import { Kanban_categories } from "@/types/supabase.tables.types";
 import useScrumBoardItemBackDropStore from "@/zustand/createScrumBoardItemStore";
 import { useMutation } from "@tanstack/react-query";
 import { toast } from "react-toastify";
-import styled from "styled-components";
+import {
+  StCreateCategoryBackDrop,
+  StCreateCategoryBackDropWrapper,
+} from "../../styles/backdrop.styles";
 
 interface Props {
   isOpen: boolean;
@@ -25,10 +28,10 @@ export default function CreateCategoryBackDrop({ isOpen }: Props) {
   });
 
   const handleOnClickSelectCategory = (category: Kanban_categories) => {
-    if (backDropType !== BACK_DROP_TYPE_CREATE) {
+    if (backDropType !== BACK_DROP_TYPE_CREATE && kanbanItem) {
       updateCategoryMutate.mutate(
         {
-          id: kanbanItem?.id!,
+          id: kanbanItem.id!,
           updateCategoryId: category.id,
         },
         {
@@ -54,55 +57,8 @@ export default function CreateCategoryBackDrop({ isOpen }: Props) {
               {category.name}
             </StCreateCategoryBackDrop>
           );
+        return null;
       })}
     </StCreateCategoryBackDropWrapper>
   );
 }
-
-const StCreateCategoryBackDropWrapper = styled.div<{
-  $isOpen: boolean;
-}>`
-  display: ${(props) => (props.$isOpen ? "flex" : "none")};
-  position: absolute;
-  bottom: 0;
-  transform: translateY(100%);
-  left: 0;
-  width: ${(props) => props.theme.unit[128]};
-  flex-direction: column;
-  border-radius: ${(props) => props.theme.border.radius[8]};
-  background-color: ${(props) => props.theme.color.bg.primary};
-  box-shadow: ${(props) => props.theme.elevation.Light.shadow8};
-  z-index: 100;
-`;
-
-const StCreateCategoryBackDrop = styled.div`
-  width: 100%;
-  height: ${(props) => props.theme.spacing[32]};
-  padding: ${(props) => props.theme.spacing[8]};
-  background-color: ${(props) => props.theme.color.bg.secondary};
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  //prettier-ignore
-  padding: ${(props) => props.theme.spacing[24]} ${(props) =>
-    props.theme.spacing[12]};
-  &:hover {
-    background-color: ${(props) => props.theme.color.bg.primary};
-    /* box-shadow: ${(props) => props.theme.elevation.Light.shadow16}; */
-  }
-  color: ${(props) => props.theme.color.text.secondary};
-  font-family: var(--point-font);
-  font-size: ${(props) => props.theme.unit[12]};
-  font-style: normal;
-  font-weight: 700;
-  line-height: 100%; /* 11px */
-  letter-spacing: -0.11px;
-  &:nth-child(1) {
-    border-radius: ${(props) =>
-      `${props.theme.border.radius["8"]} ${props.theme.border.radius["8"]} 0 0`};
-  }
-  &:nth-last-child(1) {
-    border-radius: ${(props) =>
-      `0 0 ${props.theme.border.radius["8"]} ${props.theme.border.radius["8"]}`};
-  }
-`;
