+++ nbc-nep-front/src/components/scrumboard/detail/createBackDrop/CreateInput.tsx
@@ -1,5 +1,5 @@
 import CreateCategoryInputBox from "@/components/scrumboard/detail/createBackDrop/CreateCategoryInputBox";
-import styled from "styled-components";
+import { StCreateInputBackDropWrapper } from "../../styles/backdrop.styles";
 
 export default function CreateInput() {
   return (
@@ -8,9 +8,3 @@ export default function CreateInput() {
     </StCreateInputBackDropWrapper>
   );
 }
-
-const StCreateInputBackDropWrapper = styled.div`
-  position: relative;
-  width: 275px;
-  padding: ${(props) => props.theme.spacing[4]};
-`;
