+++ nbc-nep-front/src/components/scrumboard/detail/createBackDrop/SelectAssigneesList.tsx
@@ -1,7 +1,11 @@
+import { closeIcon } from "@/assets/boards";
 import useScrumBoardMemberSearchStore from "@/zustand/scrumBoardMemberStore";
 import Image from "next/image";
-import styled from "styled-components";
-import { closeIcon } from "@/assets/boards";
+import {
+  StDeleteButton,
+  StSelectAssigneesCard,
+  StSelectAssigneesListWrapper,
+} from "../../styles/assignee.styles";
 
 interface Props {
   tagType: "assignees" | "labels";
@@ -15,14 +19,14 @@ export default function SelectAssigneesList({ tagType }: Props) {
   };
   return (
     <StSelectAssigneesListWrapper $tagType={tagType === "assignees"}>
-      {assignees.map((assignee, index) => (
-        <StSelectAssigneesCard key={index}>
+      {assignees.map((assignee) => (
+        <StSelectAssigneesCard key={assignee.created_at + assignee.id}>
           <span>{assignee.space_display_name}</span>
           {tagType === "assignees" && (
             <StDeleteButton
               onClick={() => handleDeleteAssignees(assignee.user_id!)}
             >
-              <Image src={closeIcon} alt={"삭제"} width={12} height={12} />
+              <Image src={closeIcon} alt="삭제" width={12} height={12} />
             </StDeleteButton>
           )}
         </StSelectAssigneesCard>
@@ -30,50 +34,3 @@ export default function SelectAssigneesList({ tagType }: Props) {
     </StSelectAssigneesListWrapper>
   );
 }
-
-const StSelectAssigneesListWrapper = styled.div<{ $tagType: boolean }>`
-  width: 100%;
-  border: 1px solid;
-  border-radius: ${(props) => props.theme.border.radius[8]};
-  border: ${({ $tagType }) => ($tagType ? "1px" : "0")} solid
-    ${(props) => props.theme.color.border.secondary};
-  padding: ${(props) => props.theme.spacing[8]};
-  display: flex;
-  gap: ${(props) => props.theme.spacing[8]};
-  flex-wrap: wrap;
-`;
-
-const StSelectAssigneesCard = styled.div`
-  display: flex;
-  //prettier-ignore
-  padding: ${(props) => props.theme.spacing[4]} ${(props) =>
-    props.theme.spacing[8]};
-  justify-content: center;
-  align-items: center;
-  gap: ${(props) => props.theme.spacing[4]};
-  border-radius: ${(props) => props.theme.border.radius["circle"]};
-  background: rgba(239, 246, 255, 0.8);
-
-  > span {
-    color: ${(props) => props.theme.color.blue["600"]};
-    text-overflow: ellipsis;
-    font-family: var(--sub-font);
-    font-size: ${(props) => props.theme.unit[14]};
-    font-style: normal;
-    font-weight: 400;
-    line-height: 100%;
-    letter-spacing: 0.28px;
-    text-transform: uppercase;
-  }
-`;
-
-const StDeleteButton = styled.button`
-  border: none;
-  width: 12px;
-  background-color: unset;
-  height: 12px;
-  padding: unset;
-  &:hover {
-    background-color: unset;
-  }
-`;
