+++ nbc-nep-front/src/components/scrumboard/libs/BackDropType.tsx
@@ -1,10 +1,10 @@
-import { BackDropType } from "@/types/scrum.types";
 import {
   BACK_DROP_TYPE_DETAIL,
   BACK_DROP_TYPE_UPDATE,
-} from "@/components/scrumboard/constants/constants";
+} from "@/components/scrumboard/constants";
 import CreateBackDropCtaButton from "@/components/scrumboard/detail/createBackDrop/CreateBackDropCTAButton";
 import useCreateScrumButtons from "@/hooks/scrumBoard/useCreateScrumButtons";
+import { BackDropType } from "@/types/scrum.types";
 
 export default function BackDropTypeButtonGroup(type: BackDropType) {
   const {
@@ -21,18 +21,18 @@ export default function BackDropTypeButtonGroup(type: BackDropType) {
       return (
         <>
           <CreateBackDropCtaButton
-            buttonType={"submit"}
-            forwardText={"수정"}
+            buttonType="submit"
+            forwardText="수정"
             handle={handleOnClickUpdate}
           />
           <CreateBackDropCtaButton
-            buttonType={"button"}
-            forwardText={"삭제"}
+            buttonType="button"
+            forwardText="삭제"
             handle={handleOnClickDelete}
           />
           <CreateBackDropCtaButton
-            buttonType={"button"}
-            forwardText={"닫기"}
+            buttonType="button"
+            forwardText="닫기"
             handle={handleOnClickBackDropClose}
           />
         </>
@@ -41,13 +41,13 @@ export default function BackDropTypeButtonGroup(type: BackDropType) {
       return (
         <>
           <CreateBackDropCtaButton
-            buttonType={"submit"}
-            forwardText={"저장"}
+            buttonType="submit"
+            forwardText="저장"
             handle={handleOnClickUpdateConfirm}
           />
           <CreateBackDropCtaButton
-            buttonType={"button"}
-            forwardText={"취소"}
+            buttonType="button"
+            forwardText="취소"
             handle={handleOnClickUpdateCancel}
           />
         </>
@@ -56,13 +56,13 @@ export default function BackDropTypeButtonGroup(type: BackDropType) {
       return (
         <>
           <CreateBackDropCtaButton
-            buttonType={"submit"}
-            forwardText={"생성"}
+            buttonType="submit"
+            forwardText="생성"
             handle={handleOnClickCreate}
           />
           <CreateBackDropCtaButton
-            buttonType={"button"}
-            forwardText={"닫기"}
+            buttonType="button"
+            forwardText="닫기"
             handle={handleOnClickBackDropClose}
           />
         </>
