+++ nbc-nep-front/src/components/scrumboard/styles/scrumBoardList.styles.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StScrumBoardListWrapper = styled.ul`
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: ${(props) => props.theme.spacing[24]};
+`;
